+++ nmoo/noises/__init__.py
@@ -4,3 +4,4 @@ Noise wrappers.
 __docformat__ = "google"
 
 from .gaussian import GaussianNoise
+from .uniform import UniformNoise